Fixes for TestingSyncUpForm
