--- Sources/ComposableArchitecture/Documentation.docc/Tutorials/BuildingSyncUps/03-SyncUpForm/TestingSyncUpForm-01-code-0004-previous.swift
@@ -1,11 +1,0 @@
-import ComposableArchitecture
-import XCTest
-
-@testable import SyncUps
-
-class SyncUpFormTests: XCTestCase {
-  @MainActor
-  func testRemoveAttendee() async {
-    // ...
-  }
-}
