--- Sources/ComposableArchitecture/Documentation.docc/Tutorials/BuildingSyncUps/03-SyncUpForm/TestingSyncUpForm-01-code-0004.swift
@@ -1,28 +1,0 @@
-import ComposableArchitecture
-import XCTest
-
-@testable import SyncUps
-
-class SyncUpFormTests: XCTestCase {
-  @MainActor
-  func testRemoveFocusedAttendee() async {
-    let attendee1 = Attendee(id: Attendee.ID())
-    let attendee2 = Attendee(id: Attendee.ID())
-    let store = TestStore(
-      initialState: SyncUpForm.State(
-        focus: .attendee(attendee1.id),
-        syncUp: SyncUp(
-          id: SyncUp.ID(),
-          attendees: [attendee1, attendee2]
-        )
-      )
-    ) {
-      SyncUpForm()
-    }
-  }
-
-  @MainActor
-  func testRemoveAttendee() async {
-    // ...
-  }
-}
