+++ Sources/ComposableArchitecture/Documentation.docc/Tutorials/BuildingSyncUps/03-SyncUpForm/TestingSyncUpForm-02-code-0007.swift
@@ -18,8 +18,9 @@ class SyncUpFormTests: XCTestCase {
     }
 
     await store.send(.addAttendeeButtonTapped) {
-      state.focus = .attendee(Attendee.ID(0))
-      state.syncUp.attendees.append(Attendee(id: Attendee.ID(0)))
+      let attendee = Attendee(id: Attendee.ID(0))
+      $0.focus = .attendee(attendee.id)
+      $0.syncUp.attendees.append(attendee)
     }
   }
